add delete banner

+++ models/banner.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Banner struct {
-	BannerID  uint            `json:"banner_id" gorm:"primarykey"`
-	FeatureID uint            `json:"feature_id"`
+	BannerID  uint64          `json:"banner_id" gorm:"primarykey"`
+	FeatureID uint64          `json:"feature_id"`
 	Content   json.RawMessage `json:"content" gorm:"type:json"` // For JSON type in postgres, RawMessage is []byte
 	IsActive  bool            `json:"is_active"`
 	CreatedAt time.Time       `json:"created_at" gorm:"type:timestamp"`
@@ -16,14 +16,14 @@ type Banner struct {
 }
 
 type BannerRequestBody struct {
-	TagIds    []uint          `json:"tag_ids"`
-	FeatureID uint            `json:"feature_id"`
+	TagIds    []uint64        `json:"tag_ids"`
+	FeatureID uint64          `json:"feature_id"`
 	Content   json.RawMessage `json:"content"` // For JSON type in postgres, RawMessage is []byte
 	IsActive  bool            `json:"is_active"`
 }
 
 type BannerCreatePayload struct {
-	FeatureID uint            `json:"feature_id"`
+	FeatureID uint64          `json:"feature_id"`
 	Content   json.RawMessage `json:"content"` // For JSON type in postgres, RawMessage is []byte
 	IsActive  bool            `json:"is_active"`
 	CreatedAt time.Time       `json:"created_at"`
