+++ models/tag.go
@@ -1,7 +1,7 @@
 package models
 
 type Tag struct {
-	TagID       uint   `json:"tag_id" gorm:"primarykey" `
+	TagID       uint64 `json:"tag_id" gorm:"primarykey" `
 	Description string `json:"description"`
 }
