+++ service/bannerTag.go
@@ -23,10 +23,10 @@ func (s *BannerTagService) FindAll() ([]models.BannerTag, error) {
 	return s.repo.FindAll()
 }
 
-func (s *BannerTagService) FindByID(bannerID, tagID uint) (*models.BannerTag, error) {
+func (s *BannerTagService) FindByID(bannerID, tagID uint64) (*models.BannerTag, error) {
 	return s.repo.FindByID(bannerID, tagID)
 }
 
-func (s *BannerTagService) Delete(bannerID, tagID uint) error {
+func (s *BannerTagService) Delete(bannerID, tagID uint64) error {
 	return s.repo.Delete(bannerID, tagID)
 }
